requirements.txt
# Client (used by transpiler and runtime)
# Already required by ComfyUI
aiohttp

# 1.5.9: https://github.com/erdewit/nest_asyncio/issues/87
nest_asyncio ~= 1.0, >= 1.5.9

# Runtime
# Already required by ComfyUI
Pillow

numpy